bug 1816751: rename productlib to libproduct
This fixes libproduct to be in line with current conventions.
