+++ webapp/crashstats/supersearch/views.py
@@ -17,7 +17,7 @@ from django.views.decorators.http import require_POST
 
 from django_ratelimit.decorators import ratelimit
 
-from crashstats import productlib
+from crashstats import libproduct
 from crashstats.crashstats import models, utils
 from crashstats.crashstats.decorators import track_view
 from crashstats.crashstats.utils import render_exception, urlencode_obj
@@ -60,7 +60,7 @@ def get_allowed_fields(user):
 
 def get_supersearch_form(request):
     platforms = list(models.Platform.objects.values_list("name", flat=True))
-    products = [product.name for product in productlib.get_products()]
+    products = [product.name for product in libproduct.get_products()]
 
     # FIXME(willkg): this hardcodes always getting Firefox versions which
     # seems unhelpful
